requirements.txt
gunicorn
aiofiles==24.1.0
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
html5tagger==1.3.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
multidict==6.1.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyreadline3==3.5.4
requests==2.32.3
rsa==4.9
sanic==24.6.0
sanic-routing==23.12.0
tqdm==4.66.5
tracerite==1.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
